font loader: fix x64 pointer truncation in 32-byte align mask
The crash on the magic-check read was actually downstream of an x64
pointer truncation in our align mask. `~31u` is a 32-bit constant
(0xFFFFFFE0); when AND-ed with a 64-bit uintptr_t pointer, the
upper 32 bits get zeroed. So:
  pvp = align(dp & ~31u)
turns a 0x000001B4_xxxxxxxx pointer into 0x00000000_xxxxxxxx --
deref of which AVs (the address 0x00000000A8EA6180 we saw in the
last crash report).

Fix: cast the mask itself to uintptr_t -> `~(uintptr_t)31` is
0xFFFFFFFFFFFFFFE0 -> AND preserves all 64 pointer bits.

This is the same x64-truncation issue documented in
feedback_x64_ptr_truncation.md from earlier ps2_texture.c work --
should have caught it on the first read but missed the operator
precedence. Walk diagnostic lines confirm walk arrives at file
offset 8444; align should give 8448 = the PLI/TIM2 magic spotted
by the very first scan.

+++ src/ps2/veronica/prog/system.c
@@ -1377,8 +1377,15 @@ void bhSysCallMovie()
                                 dp += 4;
                             }
                             if (blksz == -1) goto skip_font_load;
-                            /* 32-byte align (effect.c:139). */
-                            uintptr_t align = ((uintptr_t)dp + 31u) & ~31u;
+                            /* 32-byte align (effect.c:139). On x64 the
+                             * mask MUST be uintptr_t-wide -- `~31u` is
+                             * 32-bit (0xFFFFFFE0), and AND-ing it with
+                             * a 64-bit pointer zeroes the high half,
+                             * truncating the pointer to a low-32 garbage
+                             * address (caused the previous crash at
+                             * bhSetMemPvpTexture+0xc1 reading a
+                             * 0x00000000xxxxxxxx pointer). */
+                            uintptr_t align = ((uintptr_t)dp + 31u) & ~(uintptr_t)31u;
                             unsigned char* pvp = (unsigned char*)align;
                             int off = (int)(pvp - buf);
                             if (off + 32 < sz) {
